[FLINK-24017][kryo] Move FlinkChillPackageRegistrar to flink-java
