+++ flink-java/src/main/java/org/apache/flink/api/java/typeutils/runtime/kryo/FlinkChillPackageRegistrar.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.runtime.types;
+package org.apache.flink.api.java.typeutils.runtime.kryo;
 
 import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.Serializer;
